changed app to .jsx, and moved styles files into folder, and deleted reference to images in manifest.json
